backend: Add test-specific app and JSON tests
This commit adds a test-specific app called mdh_tests which contains
specific models used only for testing. It also migrates test-related
utility functions and the pep8 compliance test to this app.

This furthermore adds extensive tests for storing and loading JSON
objects to and from the database. In particular, this is done for
SmartJSONField and DumbJSONField, ensure that objects can be stored
using native JSONField()s for postgres and string-based fallbacks for
other database types.  # Please enter the commit message for your
changes. Lines starting